+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 ## Changes
+- Renamed all `OptionalData` into `AdditionalAttributes` and `OptionalDataEntry` into `Attribute` ([#422](https://github.com/desmos-labs/desmos/issues/422))
 - Capped the `OptionalDataEntry.Key` length ([#393](https://github.com/desmos-labs/desmos/issues/393)])
 - Removed `--allows-comments` in favor of `--disable-comments` ([#386](https://github.com/desmos-labs/desmos/issues/386))
 - Standardized `PostID` naming convetions ([#419](https://github.com/desmos-labs/desmos/issues/419))